renaming and cleanup; tests will break from now on
