+++ src/Ducc0.jl
@@ -1,171 +1,245 @@
+# proposed interface to pass array and type information from Julia to C++
+
+# This code does not work out of the box since I have not updated ducc0_jll yet.
+# It should mainly serve as a base for discussion.
+
+# Formatting: using JuliaFormatter; format_file("proposed_interface.jl",remove_extra_newlines=true)
+
 module Ducc0
 
-using Libdl
-using ducc0_jll
+module Support
 
-export nufft_u2nu!, nufft_nu2u!
-export NufftPlan, make_nufft_plan, delete_nufft_plan
-export planned_nu2u, planned_u2nu
+#import ducc0_jll
+#const libducc = ducc0_jll.libducc_julia
+const libducc = "/home/martin/codes/ducc/julia/ducc_julia.so" # FIXME
+#"/home/martin/codes/ducc/julia/ducc_julia.so"
 
-function nufft_u2nu!(
-    ndim::Integer,
-    npoints::Integer,
-    shape,
-    grid::Array{Float64},
-    coord::Array{Float64},
-    out::Array{Float64},
-    sigma_min,
-    sigma_max,
-    fft_order;
-    forward = true,
-    verbosity = false,
-    epsilon = 1e-7,
-    nthreads = 1,
-    periodicity = 2π
+struct ArrayDescriptor
+    shape::NTuple{10,UInt64}  # length of every axis
+    stride::NTuple{10,Int64}  # stride along every axis (in elements)
+    data::Ptr{Cvoid}          # pointer to the first array element
+    ndim::UInt8               # number of dimensions
+    dtype::UInt8              # magic values determining the data type
+end
+
+# convert data types to type codes for communication with ducc
+function typecode(tp::Type)
+    if tp <: AbstractFloat
+        return sizeof(tp(0)) - 1
+    elseif tp <: Unsigned
+        return sizeof(tp(0)) - 1 + 32
+    elseif tp <: Signed
+        return sizeof(tp(0)) - 1 + 16
+    elseif tp == Complex{Float32}
+        return typecode(Float32) + 64
+    elseif tp == Complex{Float64}
+        return typecode(Float64) + 64
+    end
+end
+
+function desc(arr::StridedArray{T,N}) where {T,N}
+    @assert N <= 10
+    ArrayDescriptor(
+        NTuple{10,UInt64}(i <= N ? size(arr)[i] : 0 for i = 1:10),
+        NTuple{10,Int64}(i <= N ? strides(arr)[i] : 0 for i = 1:10),
+        pointer(arr),
+        N,
+        typecode(T),
     )
+end
 
-    @assert length(coord) == ndim * npoints
-    @assert length(out) == 2npoints
-    
-    ccall(
-        (:nufft_u2nu_julia_double, libducc_julia),
-        Cvoid,
-        (
-            Csize_t,       # ndim
-            Csize_t,       # npoints
-            Ptr{Csize_t},  # shape
-            Ptr{Cdouble},  # grid
-            Ptr{Cdouble},  # coord
-            Cint,          # forward
-            Cdouble,       # epsilon
-            Csize_t,       # nthreads
-            Ptr{Cdouble},  # out
-            Csize_t,       # verbosity
-            Cdouble,       # sigma_min
-            Cdouble,       # sigma_max
-            Cdouble,       # periodicity
-            Cint,          # fft_order
-        ),
+Dref = Ref{ArrayDescriptor}
+
+export libducc, desc, Dref
+
+end  # module Support
+
+module Nufft
+
+using ..Support
+
+function best_epsilon(
+    ndim::Unsigned,
+    singleprec::Bool;
+    sigma_min::AbstractFloat = 1.1,
+    sigma_max::AbstractFloat = 2.6,
+)
+    res = ccall(
+        (:nufft_best_epsilon, libducc),
+        Cdouble,
+        (Csize_t, Cint, Cdouble, Cdouble),
         ndim,
-        npoints,
-        shape,
-        grid,
-        coord,
-        forward ? 1 : 0,
+        singleprec,
+        sigma_min,
+        sigma_max,
+    )
+    res <= 0 && throw(error())
+    return res
+end
+
+function u2nu!(
+    coord::StridedArray{T,2},
+    grid::StridedArray{T2,D},
+    points::StridedArray{T2,1};
+    forward::Bool = true,
+    verbose::Bool = false,
+    epsilon::AbstractFloat = 1e-5,
+    nthreads::Unsigned = UInt32(1),
+    sigma_min::AbstractFloat = 1.1,
+    sigma_max::AbstractFloat = 2.6,
+    periodicity::AbstractFloat = 2π,
+    fft_order::Bool = true,
+) where {T,T2,D}
+    res = Vector{T2}(undef, size(coord)[2])
+    GC.@preserve coord grid res
+    ret = ccall(
+        (:nufft_u2nu, libducc),
+        Cint,
+        (
+            Dref,
+            Dref,
+            Cint,
+            Cdouble,
+            Csize_t,
+            Dref,
+            Csize_t,
+            Cdouble,
+            Cdouble,
+            Cdouble,
+            Cint,
+        ),
+        desc(grid),
+        desc(coord),
+        0,
         epsilon,
         nthreads,
-        out,
-        verbosity ? 1 : 0,
+        desc(res),
+        verbose,
         sigma_min,
         sigma_max,
         periodicity,
         fft_order,
     )
+    ret != 0 && throw(error())
+    return res
 end
 
-function nufft_nu2u!(
-    ndim::Integer,
-    npoints::Integer,
-    shape,
-    points::Array{Float64},
-    coord::Array{Float64},
-    out::Array{Float64},
-    sigma_min,
-    sigma_max,
-    fft_order;
-    forward = true,
-    verbosity = false,
-    epsilon = 1e-7,
-    nthreads = 1,
-    periodicity = 2π
-    )
+function u2nu(
+    coord::StridedArray{T,2},
+    grid::StridedArray{T2,D};
+    forward::Bool = true,
+    verbose::Bool = false,
+    epsilon::AbstractFloat = 1e-5,
+    nthreads::Unsigned = UInt32(1),
+    sigma_min::AbstractFloat = 1.1,
+    sigma_max::AbstractFloat = 2.6,
+    periodicity::AbstractFloat = 2π,
+    fft_order::Bool = true,
+) where {T,T2,D}
+    res = Vector{T2}(undef, size(coord)[2])
+    return u2nu!(coord, grid, res, forward=forward, verbose=verbose, epsilon=epsilon, nthreads=nthreads, sigma_min=sigma_min, sigma_max=sigma_max, periodicity=periodicity, fft_order=fft_order)
+end
 
-    @assert length(coord) == ndim * npoints
-    @assert length(out) == 2prod(shape)
-    
-    ccall(
-        (:nufft_nu2u_julia_double, libducc_julia),
-        Cvoid,
+function nu2u!(
+    coord::StridedArray{T,2},
+    points::StridedArray{T2,1},
+    uniform::StridedArray{T2,D};
+    forward::Bool = true,
+    verbose::Bool = false,
+    epsilon::AbstractFloat = 1e-5,
+    nthreads::Unsigned = UInt32(1),
+    sigma_min::AbstractFloat = 1.1,
+    sigma_max::AbstractFloat = 2.6,
+    periodicity::AbstractFloat = 2π,
+    fft_order::Bool = true,
+) where {T,T2,D}
+    GC.@preserve coord points uniform
+    ret = ccall(
+        (:nufft_nu2u, libducc),
+        Cint,
         (
-            Csize_t,       # ndim
-            Csize_t,       # npoints
-            Ptr{Csize_t},  # shape
-            Ptr{Cdouble},  # points
-            Ptr{Cdouble},  # coord
-            Cint,          # forward
-            Cdouble,       # epsilon
-            Csize_t,       # nthreads
-            Ptr{Cdouble},  # out
-            Csize_t,       # verbosity
-            Cdouble,       # sigma_min
-            Cdouble,       # sigma_max
-            Cdouble,       # periodicity
-            Cint,          # fft_order
+            Dref,
+            Dref,
+            Cint,
+            Cdouble,
+            Csize_t,
+            Dref,
+            Csize_t,
+            Cdouble,
+            Cdouble,
+            Cdouble,
+            Cint,
         ),
-        ndim,
-        npoints,
-        shape,
-        points,
-        coord,
-        forward ? 1 : 0,
+        desc(points),
+        desc(coord),
+        0,
         epsilon,
         nthreads,
-        out,
-        verbosity ? 1 : 0,
+        desc(uniform),
+        verbose,
         sigma_min,
         sigma_max,
         periodicity,
         fft_order,
     )
+    ret != 0 && throw(error())
+    return uniform
 end
 
-
-@doc raw"""
-    struct NufftPlan
-
-A plan for Nufft. Create one using [`make_nufft_plan`](@ref) and destroy
-it using [`delete_nufft_plan`](@ref).
-"""
-struct NufftPlan
-    ptr::Ptr{Cvoid}
+function nu2u(
+    coord::StridedArray{T,2},
+    points::StridedArray{T2,1},
+    N::NTuple{D,Int};
+    forward::Bool = true,
+    verbose::Bool = false,
+    epsilon::AbstractFloat = 1e-5,
+    nthreads::Unsigned = UInt32(1),
+    sigma_min::AbstractFloat = 1.1,
+    sigma_max::AbstractFloat = 2.6,
+    periodicity::AbstractFloat = 2π,
+    fft_order::Bool = true,
+) where {T,T2,D}
+    res = Array{T2}(undef, N)
+    return nu2u!(coord, points, res, forward=forward, verbose=verbose, epsilon=epsilon, nthreads=nthreads, sigma_min=sigma_min, sigma_max=sigma_max, periodicity=periodicity, fft_order=fft_order) 
 end
 
+mutable struct NufftPlan
+    N::Vector{UInt64}
+    npoints::Int
+    cplan::Ptr{Cvoid}
+end
 
-function make_nufft_plan(
-    nu2u,
-    ndim,
-    npoints,
-    shape,
-    coord::Array{Float64},
-    sigma_min,
-    sigma_max,
-    fft_order;
-    epsilon = 1e-7,
-    nthreads = 1,
-    periodicity = 2π
-    )
+function delete_plan!(plan::NufftPlan)
+    if plan.cplan != C_NULL
+        ret = ccall((:nufft_delete_plan, libducc), Cint, (Ptr{Cvoid},), plan.cplan)
+        ret != 0 && throw(error())
+        plan.cplan = C_NULL
+    end
+end
 
-    result = ccall(
-        (:make_nufft_plan_double, libducc_julia),
+function make_plan(
+    coords::Matrix{T},
+    N::NTuple{D,Int};
+    nu2u::Bool = false,
+    epsilon::AbstractFloat = 1e-5,
+    nthreads::Unsigned = UInt32(1),
+    sigma_min::AbstractFloat = 1.1,
+    sigma_max::AbstractFloat = 2.6,
+    periodicity::AbstractFloat = 2π,
+    fft_order::Bool = true,
+) where {T,D}
+    N2 = Vector{UInt64}(undef, D)
+    for i = 1:D
+        N2[i] = N[i]
+    end
+    GC.@preserve N2 coords
+    ptr = ccall(
+        (:nufft_make_plan, libducc),
         Ptr{Cvoid},
-        (
-            Cint,         # nu2u
-            Csize_t,      # ndim
-            Csize_t,      # npoints
-            Ptr{Csize_t}, # shape
-            Ptr{Cdouble}, # coord
-            Cdouble,      # epsilon
-            Csize_t,      # nthreads
-            Cdouble,      # sigma_min
-            Cdouble,      # sigma_max
-            Cdouble,      # periodicity
-            Cint,         # fft_order
-        ),
+        (Cint, Dref, Dref, Cdouble, Csize_t, Cdouble, Cdouble, Cdouble, Cint),
         nu2u,
-        ndim,
-        npoints,
-        shape,
-        coord,
+        desc(N2),
+        desc(coords),
         epsilon,
         nthreads,
         sigma_min,
@@ -174,90 +248,263 @@ function make_nufft_plan(
         fft_order,
     )
 
-    return NufftPlan(result)
+    ptr == C_NULL && throw(error())
+    p = NufftPlan(N2, size(coords)[2], ptr)
+    finalizer(p -> begin
+        delete_plan!(p)
+    end, p)
+
+    return p
 end
 
-function delete_nufft_plan(plan::NufftPlan)
-    ccall(
-        (:delete_nufft_plan_double, libducc_julia),
-        Cvoid,
-        (Ptr{Cvoid},),
-        plan.ptr,
-    )
-end
-
-function planned_nu2u(
+function nu2u_planned!(
     plan::NufftPlan,
-    points::Array{Float64},
-    uniform::Array{Float64};
-    forward = true,
-    verbosity = false,
+    points::StridedArray{T,1},
+    uniform::StridedArray{T};
+    forward::Bool = false,
+    verbose::Bool = false,
+) where {T}
+    GC.@preserve points uniform
+    ret = ccall(
+        (:nufft_nu2u_planned, libducc),
+        Cint,
+        (Ptr{Cvoid}, Cint, Csize_t, Dref, Dref),
+        plan.cplan,
+        forward,
+        verbose,
+        desc(points),
+        desc(uniform),
     )
-    
-    ccall(
-        (:planned_nu2u, libducc_julia),
-        Cvoid,
-        (Ptr{Cvoid}, Cint, Csize_t, Ptr{Cdouble}, Ptr{Cdouble}),
-        plan.ptr,
-        forward ? 1 : 0,
-        verbosity ? 1 : 0,
-        points,
-        uniform,
-    )
+    ret != 0 && throw(error())
 end
 
-function planned_u2nu(
+function nu2u_planned(
     plan::NufftPlan,
-    points::Array{Float64},
-    uniform::Array{Float64};
-    forward = true,
-    verbosity = false,
-    )
-    
-    ccall(
-        (:planned_u2nu, libducc_julia),
-        Cvoid,
-        (Ptr{Cvoid}, Cint, Csize_t, Ptr{Cdouble}, Ptr{Cdouble}),
-        plan.ptr,
-        forward ? 1 : 0,
-        verbosity ? 1 : 0,
-        points,
-        uniform,
-    )
+    points::StridedArray{T,1};
+    forward::Bool = false,
+    verbose::Bool = false,
+) where {T}
+    res = Array{T}(undef, Tuple(i for i in plan.N))
+    nu2u_planned!(plan, points, res, forward = forward, verbose = verbose)
+    return res
 end
 
-
-################################################################################
-# Docstrings
-
-@doc raw"""
-    nufft_u2nu!(ndim, npoints, shape, grid, coord, out, sigma_min, sigma_max, fft_order; forward=true, verbosity=false, epsilon=1e-7, nthreads=1, periodicity=2π)
-"""
-nufft_u2nu!
-
-@doc raw"""
-    nufft_nu2u!(ndim, npoints, shape, points, coord, out, sigma_min, sigma_max, fft_order; forward=true, verbosity=false, epsilon=1e-7, nthreads=1, periodicity=2π)
-"""
-nufft_nu2u!
-
-@doc raw"""
-    make_nufft_plan(nu2u, ndim, npoints, shape, coord, sigma_min, sigma_max, fft_order; epsilon=1e-7, nthreads=1, periodicity=2π)
-"""
-make_nufft_plan
-
-@doc raw"""
-    delete_nufft_plan(plan)
-"""
-delete_nufft_plan
-
-@doc raw"""
-    planned_nu2u(plan, points, uniform; forward=true, verbosity=false)
-"""
-planned_nu2u
-
-@doc raw"""
-    planned_u2nu(plan, points, uniform; forward=true, verbosity=false)
-"""
-planned_u2nu
-
+function u2nu_planned!(
+    plan::NufftPlan,
+    uniform::StridedArray{T},
+    points::StridedArray{T,1};
+    forward::Bool = true,
+    verbose::Bool = false,
+) where {T}
+    GC.@preserve uniform points
+    ret = ccall(
+        (:nufft_u2nu_planned, libducc),
+        Cint,
+        (Ptr{Cvoid}, Cint, Csize_t, Dref, Dref),
+        plan.cplan,
+        forward,
+        verbose,
+        desc(uniform),
+        desc(points),
+    )
+    ret != 0 && throw(error())
 end
+
+function u2nu_planned(
+    plan::NufftPlan,
+    uniform::StridedArray{T};
+    forward::Bool = true,
+    verbose::Bool = false,
+) where {T}
+    res = Array{T}(undef, plan.npoints)
+    u2nu_planned!(plan, uniform, res, forward = forward, verbose = verbose)
+    return res
+end
+
+end  # module Nufft
+
+module Sht
+
+using ..Support
+
+function alm2leg!(
+    alm::StridedArray{Complex{T},2},
+    leg::StridedArray{Complex{T},3},
+    spin::Unsigned,
+    lmax::Unsigned,
+    mval::StridedArray{Csize_t,1},
+    mstart::StridedArray{Csize_t,1}, # FIXME: this is still 0-based at the moment
+    lstride::Int,
+    theta::StridedArray{Cdouble,1},
+    nthreads::Unsigned,
+) where {T}
+    GC.@preserve alm mval mstart theta leg begin
+        ret = ccall(
+            (:sht_alm2leg, libducc),
+            Cint,
+            (Dref, Csize_t, Csize_t, Dref, Dref, Cptrdiff_t, Dref, Csize_t, Dref),
+            desc(alm),
+            spin,
+            lmax,
+            desc(mval),
+            desc(mstart),
+            lstride,
+            desc(theta),
+            nthreads,
+            desc(leg),
+        )
+    end
+    ret != 0 && throw(error())
+    return leg
+end
+
+function alm2leg(
+    alm::StridedArray{Complex{T},2},
+    spin::Unsigned,
+    lmax::Unsigned,
+    mval::StridedArray{Csize_t,1},
+    mstart::StridedArray{Csize_t,1}, # FIXME: this is still 0-based at the moment
+    lstride::Int,
+    theta::StridedArray{Cdouble,1},
+    nthreads::Unsigned,
+) where {T}
+    ncomp = size(alm, 2)
+    ntheta = length(theta)
+    nm = length(mval)
+    leg = Array{Complex{T}}(undef, (nm, ntheta, ncomp))
+    alm2leg!(alm, leg, spin, lmax, mval, mstart, lstride, theta, nthreads)
+    return leg
+end
+
+function leg2alm!(
+    leg::StridedArray{Complex{T},3},
+    alm::StridedArray{Complex{T},2},
+    spin::Unsigned,
+    lmax::Unsigned,
+    mval::StridedArray{Csize_t,1},
+    mstart::StridedArray{Csize_t,1}, # FIXME: this is still 0-based at the moment
+    lstride::Int,
+    theta::StridedArray{Cdouble,1},
+    nthreads::Unsigned,
+) where {T}
+    GC.@preserve leg mval mstart theta alm begin
+        ret = ccall(
+            (:sht_leg2alm, libducc),
+            Cint,
+            (Dref, Csize_t, Csize_t, Dref, Dref, Cptrdiff_t, Dref, Csize_t, Dref),
+            desc(leg),
+            spin,
+            lmax,
+            desc(mval),
+            desc(mstart),
+            lstride,
+            desc(theta),
+            nthreads,
+            desc(alm),
+        )
+    end
+    ret != 0 && throw(error())
+    return alm
+end
+
+function leg2alm(
+    leg::StridedArray{Complex{T},3},
+    spin::Unsigned,
+    lmax::Unsigned,
+    mval::StridedArray{Csize_t,1},
+    mstart::StridedArray{Csize_t,1}, # FIXME: this is still 0-based at the moment
+    lstride::Int,
+    theta::StridedArray{Cdouble,1},
+    nthreads::Unsigned,
+) where {T}
+    ncomp = size(leg, 3)
+    alm = Array{Complex{T}}(undef, (maximum(mstart) + lmax + 1, ncomp)) # FIXME: still 0-based as well!!
+    leg2alm!(leg, alm, spin, lmax, mval, mstart, lstride, theta, nthreads)
+end
+
+function leg2map!(
+    leg::StridedArray{Complex{T},3},
+    map::StridedArray{T,2},
+    nphi::StridedArray{Csize_t,1},
+    phi0::StridedArray{Cdouble,1},
+    ringstart::StridedArray{Csize_t,1}, # FIXME: this is still 0-based at the moment
+    pixstride::Int,
+    nthreads::Unsigned,
+) where {T}
+    GC.@preserve leg nphi phi0 ringstart map begin
+        ret = ccall(
+            (:sht_leg2map, libducc),
+            Cint,
+            (Dref, Dref, Dref, Dref, Cptrdiff_t, Csize_t, Dref),
+            desc(leg),
+            desc(nphi),
+            desc(phi0),
+            desc(ringstart),
+            pixstride,
+            nthreads,
+            desc(map),
+        )
+    end
+    ret != 0 && throw(error())
+    return map
+end
+
+function leg2map(
+    leg::StridedArray{Complex{T},3},
+    nphi::StridedArray{Csize_t,1},
+    phi0::StridedArray{Cdouble,1},
+    ringstart::StridedArray{Csize_t,1}, # FIXME: this is still 0-based at the moment
+    pixstride::Int,
+    nthreads::Unsigned,
+) where {T}
+    ncomp = size(leg, 3)
+    npix = maximum(ringstart + nphi)
+    map = Array{T}(undef, (npix, ncomp))
+    leg2map!(leg, map, nphi, phi0, ringstart, pixstride, nthreads)
+end
+
+function map2leg!(
+    map::StridedArray{T,2},
+    leg::StridedArray{Complex{T},3},
+    nphi::StridedArray{Csize_t,1},
+    phi0::StridedArray{Cdouble,1},
+    ringstart::StridedArray{Csize_t,1}, # FIXME: this is still 0-based at the moment
+    pixstride::Int,
+    nthreads::Unsigned,
+) where {T}
+    GC.@preserve map nphi phi0 ringstart leg begin
+        ret = ccall(
+            (:sht_map2leg, libducc),
+            Cint,
+            (Dref, Dref, Dref, Dref, Cptrdiff_t, Csize_t, Dref),
+            desc(map),
+            desc(nphi),
+            desc(phi0),
+            desc(ringstart),
+            pixstride,
+            nthreads,
+            desc(leg),
+        )
+    end
+    ret != 0 && throw(error())
+    return leg
+end
+
+function map2leg(
+    map::StridedArray{T,2},
+    nphi::StridedArray{Csize_t,1},
+    phi0::StridedArray{Cdouble,1},
+    ringstart::StridedArray{Csize_t,1}, # FIXME: this is still 0-based at the moment
+    nm::Int,
+    pixstride::Int,
+    nthreads::Unsigned,
+) where {T}
+    ncomp = size(map, 2)
+    ntheta = length(ringstart)
+    leg = Array{Complex{T}}(undef, (nm, ntheta, ncomp))
+    map2leg!(map, leg, nphi, phi0, ringstart, pixstride, nthreads)
+end
+
+end  # module Sht
+
+end  # module Ducc0
