+++ test/runtests.jl
@@ -1,29 +1,30 @@
-using Ducc0
+include("/home/martin/codes/Ducc0.jl/src/Ducc0.jl")
+using .Ducc0.Nufft: u2nu!, nu2u!, make_plan, delete_plan!, u2nu_planned, nu2u_planned
 using Test
 
 @testset "Nufft" begin
-    let ndim = 3, npoints = 10_000_000, shape = [100, 200, 300]
-        grid = Array{Float64}(undef, prod(shape) * 2)
-        coord = Array{Float64}(undef, ndim * npoints)
-        points = Array{Float64}(undef, npoints * 2)
+    let ndim = 3, npoints = 10_000_000, shape = (100, 200, 300)
+        grid = Array{Complex{Float64}}(undef, shape)
+        coord = Array{Float64}(undef, (ndim, npoints))
+        points = Array{Complex{Float64}}(undef, (npoints,))
 
-        nufft_u2nu!(ndim, npoints, shape, grid, coord, points, 1.1, 2.5, 0, forward = true, verbosity = true, epsilon = 1e-5, nthreads=4)
-        nufft_nu2u!(ndim, npoints, shape, points, coord, grid, 1.1, 2.5, 0, forward = true, verbosity = true, epsilon = 1e-5, nthreads=4)
+        u2nu!(coord, grid, points, sigma_min=1.1, sigma_max=2.5, forward = true, verbose = true, epsilon = 1e-5, nthreads=Unsigned(4))
+        nu2u!(coord, points, grid, sigma_min=1.1, sigma_max=2.5, forward = true, verbose = true, epsilon = 1e-5, nthreads=Unsigned(4))
     end
 end
 
 @testset "Nufft.plan" begin
-    let ndim = 3, npoints = 10_000_000, shape = [100, 200, 300]
-        grid = Array{Float64}(undef, prod(shape) * 2)
-        coord = Array{Float64}(undef, ndim * npoints)
-        points = Array{Float64}(undef, npoints * 2)
+    let ndim = 3, npoints = 10_000_000, shape = (100, 200, 300)
+        grid = Array{Complex{Float64}}(undef, shape)
+        coord = Array{Float64}(undef, (ndim, npoints))
+        points = Array{Complex{Float64}}(undef, (npoints,))
 
-        plan = make_nufft_plan(0, ndim, npoints, shape, coord, 1.1, 2.6, 0, epsilon=1e-5, nthreads=4, periodicity=2π)
-        planned_u2nu(plan, grid, points, forward=true, verbosity=true)
-        delete_nufft_plan(plan)
+        plan = make_plan(coord, shape, sigma_min=1.1, sigma_max=2.6, epsilon=1e-5, nthreads=Unsigned(4), periodicity=2π)
+        u2nu_planned(plan, grid, forward=true, verbose=true)
+        delete_plan!(plan)
         
-        plan = make_nufft_plan(1, ndim, npoints, shape, coord, 1.1, 2.6, 0, epsilon=1e-5, nthreads=4, periodicity=2π)
-        planned_nu2u(plan, points, grid, forward=true, verbosity=true)
-        delete_nufft_plan(plan)
+        plan = make_plan(coord, shape, sigma_min=1.1, sigma_max=2.6, epsilon=1e-5, nthreads=Unsigned(4), periodicity=2π)
+        nu2u_planned(plan, points, forward=true, verbose=true)
+        delete_plan!(plan)
     end
 end
